Switch to `pkg` script // New OS variable
The switch to `pkg` as the packaging script enables source code
repositories to contain a build/ directory (usually for autotools and
supporting toolchain) without having to override the path to exec when
running the Docker image.

The new `$OS` variable allows packaging scripts to use the OS name (i.e.
"ubuntu" or "centos") in directory structures, without overloading
`$FLAVOR`, which still means type-of-package-created.